Add getOnClickListener to AuthenticationButton
