+++ connect/src/com/telenor/connect/ui/AuthenticationButton.java
@@ -2,6 +2,8 @@ package com.telenor.connect.ui;
 
 import com.telenor.connect.id.Claims;
 
+import android.view.View;
+
 import java.util.ArrayList;
 import java.util.Map;
 
@@ -17,6 +19,8 @@ public interface AuthenticationButton {
     Claims getClaims();
 
     int getCustomLoadingLayout();
+
+    View.OnClickListener getOnClickListener();
 
     void setAcrValues(String... acrValues);
